requirements.txt
scikit-surprise
numpy
pandas
nltk
streamlit
scikit-learn
pillow
